+++ yoko-rmi-impl/src/main/java/org/apache/yoko/rmi/impl/FieldDescriptor.java
@@ -18,11 +18,8 @@
 package org.apache.yoko.rmi.impl;
 
 import org.apache.yoko.util.concurrent.LazyReference;
-import org.omg.CORBA.MARSHAL;
-import org.omg.CORBA.SystemException;
 import org.omg.CORBA.ValueMember;
 
-import javax.rmi.PortableRemoteObject;
 import java.io.Externalizable;
 import java.io.IOException;
 import java.io.ObjectStreamField;
@@ -43,7 +40,6 @@ import static java.util.Collections.unmodifiableMap;
 import static java.util.logging.Level.FINER;
 import static org.apache.yoko.util.Exceptions.as;
 import static org.apache.yoko.util.PrivilegedActions.exAction;
-import static org.apache.yoko.util.yasf.Yasf.NON_SERIALIZABLE_FIELD_IS_ABSTRACT_VALUE;
 
 abstract class FieldDescriptor extends ModelElement implements Comparable<FieldDescriptor> {
     static Logger logger = Logger.getLogger(FieldDescriptor.class.getName());
@@ -261,783 +257,3 @@ abstract class FieldDescriptor extends ModelElement implements Comparable<FieldD
     }
 }
 
-class RemoteFieldDescriptor extends FieldDescriptor {
-    final Class<?> interfaceType;
-
-    RemoteFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-        interfaceType = (type.isInterface() ? type : findInterfaceType(type));
-    }
-
-    private static Class<?> findInterfaceType(Class<?> type) {
-        Class<?> t = type;
-
-        while (!Object.class.equals(t)) {
-            for (Class<?> intf : t.getInterfaces()) {
-                if (Remote.class.isAssignableFrom(intf)) {
-                    return intf;
-                }
-            }
-            t = t.getSuperclass();
-        }
-        throw new RuntimeException("cannot find remote interface for " + type);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        Object value = reader.readRemoteObject(interfaceType);
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeRemoteObject(getFieldContents(obj));
-    }
-
-    void copyState(final Object orig, final Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, state.copy(getFieldContents(orig)));
-        } catch (CopyRecursionException e) {
-            state.registerRecursion(new CopyRecursionResolver(orig) {
-                public void resolve(Object value) {
-                    try {
-                        setFieldContents(copy, value);
-                    } catch (IOException ex) {
-                        throw as(InternalError::new, ex, ex.getMessage());
-                    }
-                }
-            });
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) {
-        Remote value = reader.readRemoteObject(interfaceType);
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Remote value = (Remote) map.get(java_name);
-        writer.writeRemoteObject(value);
-    }
-}
-
-class AnyFieldDescriptor extends FieldDescriptor {
-    static final Logger logger = Logger.getLogger(AnyFieldDescriptor.class
-            .getName());
-
-    boolean narrowValue;
-
-    AnyFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-        narrowValue = Remote.class.isAssignableFrom(type);
-    }
-
-    public void read(ObjectReader reader, Object obj)
-            throws IOException {
-        try {
-            Object val = reader.readAny();
-            if (narrowValue && val != null && !type.isInstance(val)) {
-                try {
-                    val = PortableRemoteObject.narrow(val, this.type);
-                } catch (SecurityException ex) {
-                    logger.finer(() -> "Narrow failed" + "\n" + ex);
-                    throw ex;
-                }
-            } else if (val != null && !type.isInstance(val)) {
-                throw new MARSHAL("value is instance of "
-                        + val.getClass().getName() + " -- should be: "
-                        + type.getName());
-            }
-
-            setFieldContents(obj, val);
-        } catch (IllegalStateException ex) {
-            throw as(MARSHAL::new, ex, ex.getMessage());
-        }
-    }
-
-    public void write(ObjectWriter writer, Object obj)
-            throws IOException {
-        writer.writeAny(getFieldContents(obj));
-    }
-
-    void copyState(final Object orig, final Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, state.copy(getFieldContents(orig)));
-        } catch (CopyRecursionException e) {
-            state.registerRecursion(new CopyRecursionResolver(orig) {
-                public void resolve(Object value) {
-                    try {
-                        setFieldContents(copy, value);
-                    } catch (IOException ex) {
-                        throw as(InternalError::new, ex, ex.getMessage());
-                    }
-                }
-            });
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) {
-        Object value = reader.readAny();
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Object value = map.get(java_name);
-        writer.writeAny(value);
-    }
-}
-
-class ValueFieldDescriptor extends FieldDescriptor {
-    ValueFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        try {
-            Object value;
-            if (NON_SERIALIZABLE_FIELD_IS_ABSTRACT_VALUE.isSupported()
-                    || type.isInterface()
-                    || Serializable.class.isAssignableFrom(type)) {
-                value = reader.readValueObject(getType());
-            } else {
-                // older versions of Yoko treat non-serializable classes as abstract objects
-                value = reader.readAbstractObject();
-            }
-            setFieldContents(obj, value);
-        } catch (IllegalStateException ex) {
-            throw as(MARSHAL::new, ex, ex.getMessage());
-        }
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        if (NON_SERIALIZABLE_FIELD_IS_ABSTRACT_VALUE.isSupported()
-                || type.isInterface()
-                || Serializable.class.isAssignableFrom(type)) {
-            try {
-                writer.writeValueObject(getFieldContents(obj));
-            } catch (SystemException e) {
-                throw e;
-            } catch (Exception e) {
-                throw as(MARSHAL::new, e, "Object of class " + obj.getClass().getName() + " is not a valuetype");
-            }
-        } else {
-            // older versions of Yoko treat non-serializable classes as abstract objects
-            writer.writeObject(getFieldContents(obj));
-        }
-    }
-
-    void copyState(final Object orig, final Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, state.copy(getFieldContents(orig)));
-        } catch (CopyRecursionException e) {
-            state.registerRecursion(new CopyRecursionResolver(orig) {
-                public void resolve(Object value) {
-                    try {
-                        setFieldContents(copy, value);
-                    } catch (IOException ex) {
-                        throw as(InternalError::new, ex, ex.getMessage());
-                    }
-                }
-            });
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) {
-        Serializable value = (Serializable) reader
-                .readValueObject();
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Serializable value = (Serializable) map
-                .get(java_name);
-        writer.writeValueObject(value);
-    }
-}
-
-class StringFieldDescriptor extends FieldDescriptor {
-    StringFieldDescriptor(Class<?> owner, Class<?> type, String name,
-            Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        String value = (String) reader.readValueObject();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeValueObject(getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) {
-        String value = (String) reader.readValueObject();
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        String value = (String) map.get(java_name);
-        writer.writeValueObject(value);
-    }
-}
-
-class ObjectFieldDescriptor extends FieldDescriptor {
-    ObjectFieldDescriptor(Class<?> owner, Class<?> type, String name,
-            Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        try {
-            Object value = reader.readAbstractObject();
-            setFieldContents(obj, value);
-        } catch (IllegalStateException ex) {
-            throw as(MARSHAL::new, ex, ex.getMessage());
-        }
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeObject(getFieldContents(obj));
-    }
-
-    void copyState(final Object orig, final Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, state.copy(getFieldContents(orig)));
-        } catch (CopyRecursionException e) {
-            state.registerRecursion(new CopyRecursionResolver(orig) {
-                public void resolve(Object value) {
-                    try {
-                        setFieldContents(copy, value);
-                    } catch (IOException ex) {
-                        throw as(InternalError::new, ex, ex.getMessage());
-                    }
-                }
-            });
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) {
-        Object value = reader.readAbstractObject();
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Object value = map.get(java_name);
-        writer.writeObject(value);
-    }
-}
-
-class BooleanFieldDescriptor extends FieldDescriptor {
-    BooleanFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        boolean value = reader.readBoolean();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeBoolean((Boolean) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        map.put(java_name, reader.readBoolean());
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Boolean value = (Boolean) map.get(java_name);
-        if (value == null) {
-            writer.writeBoolean(false);
-        } else {
-            writer.writeBoolean(value);
-        }
-    }
-}
-
-class ByteFieldDescriptor extends FieldDescriptor {
-    ByteFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        byte value = reader.readByte();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeByte((Byte) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        map.put(java_name, reader.readByte());
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Byte value = (Byte) map.get(java_name);
-        if (value == null) {
-            writer.writeByte(0);
-        } else {
-            writer.writeByte(value);
-        }
-    }
-}
-
-class ShortFieldDescriptor extends FieldDescriptor {
-    ShortFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        short value = reader.readShort();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeShort((Short) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        map.put(java_name, reader.readShort());
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Short value = (Short) map.get(java_name);
-        if (value == null) {
-            writer.writeShort(0);
-        } else {
-            writer.writeShort(value);
-        }
-    }
-}
-
-class CharFieldDescriptor extends FieldDescriptor {
-    CharFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        char value = reader.readChar();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeChar((Character) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            char ch = (Character) getFieldContents(val);
-            pw.print(ch);
-            pw.print('(');
-            pw.print(Integer.toHexString(0xffff & ((int) ch)));
-            pw.print(')');
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        map.put(java_name, reader.readChar());
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Character value = (Character) map.get(java_name);
-        if (value == null) {
-            writer.writeChar(0);
-        } else {
-            writer.writeChar(value);
-        }
-    }
-}
-
-class IntFieldDescriptor extends FieldDescriptor {
-    IntFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        int value = reader.readInt();
-        logger.finest(() -> "Read int field value " + value);
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeInt((Integer) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        map.put(java_name, reader.readInt());
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Integer value = (Integer) map.get(java_name);
-        if (value == null) {
-            writer.writeInt(0);
-        } else {
-            writer.writeInt(value);
-        }
-    }
-}
-
-class LongFieldDescriptor extends FieldDescriptor {
-    LongFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        long value = reader.readLong();
-        logger.finest(() -> "Read long field value " + value);
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeLong((Long) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        map.put(java_name, reader.readLong());
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Long value = (Long) map.get(java_name);
-        if (value == null) {
-            writer.writeLong(0);
-        } else {
-            writer.writeLong(value);
-        }
-    }
-}
-
-class FloatFieldDescriptor extends FieldDescriptor {
-    FloatFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        float value = reader.readFloat();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeFloat((Float) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        Float value = reader.readFloat();
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Float value = (Float) map.get(java_name);
-        if (value == null) {
-            writer.writeFloat(0.0F);
-        } else {
-            writer.writeFloat(value);
-        }
-    }
-
-}
-
-class DoubleFieldDescriptor extends FieldDescriptor {
-    DoubleFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f, TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    public void read(ObjectReader reader, Object obj) throws IOException {
-        double value = reader.readDouble();
-        setFieldContents(obj, value);
-    }
-
-    public void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeDouble((Double) getFieldContents(obj));
-    }
-
-    void copyState(Object orig, Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, getFieldContents(orig));
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void print(PrintWriter pw, Map<Object, Integer> recurse, Object val) {
-        try {
-            pw.print(java_name);
-            pw.print("=");
-            pw.print(getFieldContents(val));
-        } catch (IllegalStateException | IOException ex) {
-            pw.print("<non-local>");
-        }
-    }
-
-    /**
-     * @see FieldDescriptor#readFieldIntoMap(ObjectReader, Map)
-     */
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) throws IOException {
-        Double value = reader.readDouble();
-        map.put(java_name, value);
-    }
-
-    /**
-     * @see FieldDescriptor#writeFieldFromMap(ObjectWriter, Map)
-     */
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        Double value = (Double) map.get(java_name);
-        if (value == null) {
-            writer.writeDouble(0.0D);
-        } else {
-            writer.writeDouble(value);
-        }
-    }
-
-}
-
-class CorbaObjectFieldDescriptor extends FieldDescriptor {
-
-    protected CorbaObjectFieldDescriptor(Class<?> owner, Class<?> type, String name, Field f,TypeRepository repository) {
-        super(owner, type, name, f, repository);
-    }
-
-    void copyState(final Object orig, final Object copy, CopyState state) {
-        try {
-            setFieldContents(copy, state.copy(getFieldContents(orig)));
-        } catch (CopyRecursionException e) {
-            state.registerRecursion(new CopyRecursionResolver(orig) {
-                public void resolve(Object value) {
-                    try {
-                        setFieldContents(copy, value);
-                    } catch (IOException ex) {
-                        throw as(InternalError::new, ex, ex.getMessage());
-                    }
-                }
-            });
-        } catch (IOException ex) {
-            throw as(InternalError::new, ex, ex.getMessage());
-        }
-    }
-
-    void read(ObjectReader reader, Object obj) throws IOException {
-        Object value = reader.readCorbaObject(type);
-        setFieldContents(obj, value);
-    }
-
-    void readFieldIntoMap(ObjectReader reader, Map<String, Object> map) {
-        Object value = reader.readCorbaObject(type);
-        map.put(java_name, value);
-
-    }
-
-    void write(ObjectWriter writer, Object obj) throws IOException {
-        writer.writeCorbaObject(getFieldContents(obj));
-    }
-
-    void writeFieldFromMap(ObjectWriter writer, Map<String, Object> map) throws IOException {
-        org.omg.CORBA.Object value = (org.omg.CORBA.Object) map.get(java_name);
-        writer.writeCorbaObject(value);
-    }
-}
